Add a mechansim for optimizing stream `read` functions.
The wasi-io input-stream function `read()` returns a `list<u8>` by
value. In practice, that means that it has to call `cabi_realloc` on
every call to allocate memory for the list, and then copy the data
from that memory into the user's buffer.

This patch adds a mechanism for defining components with a
`cabi_stream_buffer` function, which can be used as an alternative
to `cabi_realloc` in canonical options.

`cabi_stream_buffer` returns the value of a global named
`CABI_STREAM_BUFFER_PTR`, which bindings can use to store
user-supplied buffers in. This allows them to avoid any
`cabi_realloc` call.

This is similar to an optimization done in the
preview1-to-preview2 adapter.