Disallow returns-of-borrows in WIT  (#1469)
* Refactor how spans are held in `UnresolvedPackage`

Put them in a `struct` with named fields and group them together by
relevance.

* Disallow returns-of-borrows in WIT

Currently in the component model it's not valid to have a function that
returns a `borrow` resource. In WIT, however, there's no protection
currently against this until a component is actually made. Bindings
generators additionally might panic and/or assume that they won't need
to generate bindings for borrows.

This commit implements a `Resolve`-level validation that a function
never returns a `borrow<T>` transitively within it. This ensures that
the error for returning a borrowed resource shows up earlier in the
development cycle and additionally ensures that bindings generators
don't have to each specifically handle this case.

Closes bytecodealliance/wit-bindgen#907

* Fix test expectation