fix bug with libdl.so and unused libraries (#1482)
My earlier PR had a bug that led to an error when `--use-built-in-libdl` was
specified and there were unused libraries provided by the user _and_
`--stub-missing-functions` was _not_ used.  This adds a test to exercise that
case and moves the `self.use_built_in_libdl = false;` statement so that it is
always run prior to a recursive call to `Linker::encode`.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>