Implement the table64 extension to the memory64 proposal
This commit implements validation, fuzzing, printing, etc, for the
extension of the memory64 proposal that adds 64-bit tables. These tables
work similarly to memories where the index type is statically listed and
is required for all various operations. The bulk of the changes here
were to wasm-smith to generate modules with these new table
instructions.