Add the concept of a "known custom section" to wasmparser
There are a number of places throughout wasm-tools which process custom
sections, and as surfaced in #1540 they're not necessarily all handled
uniformly due to the flexible nature of custom sections. This commit aims to
improve the parsing of custom sections in the future by having a
`wasmparser::KnownCustom` enum which statically enumerates all custom
sections that are known to wasmparser and can be parsed. This
additionally provides a nice convenience for creating the typed parsers
as well.