threads: add `global.atomic.rmw.*` instructions (#1563)
* threads: add `global.atomic.rmw.*` instructions

This change adds support for the various `global.atomic.rmw.*`
instructions that are now available as a part of the
shared-everything-threads [proposal].

[proposal]: https://github.com/WebAssembly/shared-everything-threads

* Fix operand validation; refactor validation helpers

The `global.atomic.rmw.*` instructions have different validation
constraints for different instructions. `xchg` maches
`global.atomic.get|set` in that it accepts subtypes of `anyref`. But
`cmpxchg` accepts subtypes of `eqref`. All the instructions accept `i32`
and `i64`. To make this more clear (and avoid mistakes like the ones I
made), I moved these "special" cases up a level to be more visible. This
allowed removing some of the other helper functions.