WIT: Implement @since and @unstable annotations (#1508)
* Add a span to all types when parsing

* Preserve error context when highlighting errors

Previously the entire error message was replaced, losing any attached
context. This commit updates error highlighting to only augment the
single error found in the chain that's being highlighted (in the most
common case). This required a small refactoring of the `Error` type and
changes all existing users to a method-based constructor rather than
explicit struct-based construction.

* Add initial parsing of attributes

Nothing uses the results of parsing yet, that's going to come in a
future commit.

* Record spans for all types in `UnresolvedPackage`

Will be used for errors in a future commit.

* Push stability attributes into top-level AST

This commit pushes stability attributes through the resolution process
to the next stage of AST. The top-level user-facing types in
`wit-parser` now have `Stability` annotations were they can be added.

This commit notably changes the `WorldItem::Interface` enum variant to
contain a stability attribute in addition to the id listed.

* Filter out `@unstable` items that aren't enabled

This finishes support for `@unstable` and `@since` in `Resolve` by
handling all items there and specifically filtering out any disabled
items.

* Add CLI support for WIT features

* Implement printing WIT stability attributes

* Round-trip stability through the wasm binary format

This involved a number of refactorings and "tricks" to get this to work
out. Namely when possible the old format of the custom section is still
emitted to ensure older/newer tools can interoperate when possible.

* Fix compile

* Fix some compile warnings