Update the spec test suite submodule (#1576)
* Update the spec test suite submodule

This commit updates the spec test suite submodule and notably
implements the extended syntax in the text format for quoted
identifiers. Names such as `$"foo"` are not valid and equivalent to
`$foo`. This doesn't yet update any `wasmprinter`-based printing to use
this new syntax, it's just parsed to pass the various spec tests added.

This additionally updates the lexer to have a first-class notion of an
annotation instead of being lumped in with "reserved" tokens to avoid
having to deal with things like `@"foo""bar"` which is a single reserved
token but technically not an annotation. This is now modelled
differently within the lexer as `TokenKind::Annotation` or
`TokenKind::Reserved`. (and can have a better first-class error in the
future).

* Fix warning on nightly

* Clarify difference in errors