Improve error message for missing delimiter at EOF
Due to the way source maps worked the span would point to the wrong file
if the marker was pointing at the end of a file. While this still isn't
necessarily a perfect error message it should be a significant
improvement over before by actually pointing to the incorrect file.

Closes #1583