Don't recover printing malformed custom sections
Previously `wasmprinter` would buffer up the failed print, discard it,
and then print the custom section as its raw bytes. This won't be easy
if wasmprinter is transitioned to `io::Write` or a stream-based
underlying primitive because output can't easily be undone or buffered
per se.

This commit changes the custom section printing to handle binary errors.
The main print still does not fail (the module keeps getting printed)
but the raw contents of the section are no longer printed. Instead just
a comment is left, parentheses are closed, and printing continues.