Split destination for printing out of `Printer`
This commit refactors the `wasmprinter` API to split the destination of
what's to be printed out of the configuration of the printing itself.
This doesn't make a ton of sense given the exact API `wasmprinter` has
today but it'll make a future insertion of a trait object much easier.
For now though otherwise `wasmprinter::Config` is a new structure which
encapsulates configuration settings and `wasmprinter::Printer` is now
only a temporary structure used during printing.