Add color by default to printing the text format (#1596)
* Add color by default to printing the text format

This commit builds on #1593 to provide colorized output of `wasm-tools
print`, or any other subcommand that prints the text format, by default.
Color settings are controlled on the CLI with `--color` and are
automatically disabled in various situations such as outputting to a
file or when an output pipe is used.

Internally `wasmprinter` has grown a new trait, `Print`, through which
all output goes through. There are adapters for this trait for
`std::fmt::Write`, `std::io::Write`, and `termcolor::WriteColor`. Custom
trait implementations are also possible if necessary.

While here I've additionally moved custom custom section printing to the
trait itself instead of being part of the printer configuration as it
felt more appropriate as a per-sink configuration.

* Improve float colors (use comment colors for comments)