Minor updates to the playground (#1605)
* Fix hyperlink between parse/print online pages

Use a relative link instead of an absolute one to handle hosting on
different roots

* Rename playground jobs

Make it a bit more descriptive which workflow they're attached to with
just the name of the job

* Remove generated `src/bindings.rs`

Exclude the crate from testing to avoid building it on other targets.

* Remove unused `[profile.release]` section

Inherit the default settings for now and if distribution of the wasm is
a problem it can be re-added on a per-CI-job-basis.

* Exclude the playground from `cargo doc`

* Try to fill in missing file for rustfmt

* Try a blank line in a new file

* Try printf instead