Use quoted identifiers in `wasmprinter` by default (#1615)
* Use quoted identifiers in `wasmprinter` by default

This commit updates `wasmprinter` to use quoted identifiers of the form
`$"foo"` when necessary instead of synthesizing identifiers such as
`$#func4<foo>`. This helps produce more readable modules by default when
names are synthesized since if a name is unique but otherwise has
non-identifier characters then the quoted string form can be used.

While here I've additionally changed the way that non-printable
characters in strings are printed to using `\u{XXX}` syntax instead of
`\NN` syntax. This makes it a bit more obvious in unicode contexts that
a single character is present and not multiple.

* Fix some test expectations

* Migrate a number of `wasmprinter` unit tests to `tests/cli/*.wat`

This makes them easier to update as the output changes over time and
additionally easier to add new files here too.