Add `lann/wasm-wave` to crates  (#1606)
* import wasm-wave 0.6.0

with minor fixes to fit in workspace, delete dependency on wasmtime, and fix some benign warnings

Co-authored-by: Lann Martin <lann.martin@fermyon.com>

* import tests from wasm-wave 0.6.0

wasmtime test removed from these sources, since that depends on wasm_wave::wasmtime which was not imported here.

Co-authored-by: Lann Martin <lann.martin@fermyon.com>

* wave ui tests: upgrade snapbox, depend on tryfn, use new syntax

* wasm-wave: remove wasmtime test dep, and dead test code

* wasm-wave: latest wit-parser api changes

* wasm-wave: remove use of wasmtime from rustdocs

* fixes for rebase

* values and types: derive Eq in addition to PartialEq

This is due to a bug that only occurs in MSRV (1.76.0) where, for consts such as Ty::BOOL be used in patterns, the type must derive not just PartialEq but also Eq.

This restriction is relaxed in later rust releases for consts constructed directly rather than with const fns.

* replace snapbox with by-hand implementation

snapbox had a pretty trivial incompatibility with wasm32-wasi, and once i fixed that I found one of its deps did as well, so rather than shave all those yaks just replace the bit we need with this trivial code

* readme: fix url to wit

* url fixes

* delete duplicate license

* add wasm-wave to crates to publish

* add comment of todo

* add todo in comment

* gitattributes: wave test input and output are text but newlines should not be messed with

---------

Co-authored-by: Lann Martin <lann.martin@fermyon.com>