Limit component model flags to 32
This commit is an attempt to explore the design space of
WebAssembly/component-model#370. This limits, by default, the number of
flags in a component model type to 32 by default. The hope of this issue
is to be able to ratchet the maximum number of flags to make it easier
on bindings generators to not need to work with arbitrary numbers of
flags. The secondary hope is that we can ratchet flags straight to 32
instead of 64 due to it being unlikely that more than 32 flags are in
use. Once this percolates there can then be a separate feature for
enabling 33-64 flags.