wit-component: Allow `cabi_realloc` to be weakly defined (#1625)
If a weak "default" visibility symbol is defined in a shared library,
wasm-ld emits an import for the symbol for the "env" module for the case
where the symbol is overridden by other modules.

`cabi_realloc` is usually defined weakly by wit-bindgen'ed code, and if
the code is compiled with `-fvisibility=default`, it emits an import for
"env.cabi_realloc". The import leads re-exporting the symbol to the
"env" module, and it conflicts with the special re-exporting logic for
`cabi_realloc`.

This commit fixes the duplicate re-exporting issue.