Handle multiple versions during `wasm-tools component wit` (#1665)
* Handle multiple versions during `wasm-tools component wit`

This commits makes sure that if a component depends on multiple versions
of a package that it is handled correctly. It fixes the behavior by
appending the version string in case there are multiple packages with
the same name and namespaces specifically when there is only one
namespace for the given package name.

Signed-off-by: karthik2804 <karthik.ganeshram@fermyon.com>

* add tests

Signed-off-by: karthik2804 <karthik.ganeshram@fermyon.com>

---------

Signed-off-by: karthik2804 <karthik.ganeshram@fermyon.com>