Better handle stability attributes when merging WIT packages (#1669)
This commit updates the logic when merging WIT packages together to
better handle stability. This isn't perfect but it should suffice to fix
the issue in #1666. The general idea is that stability is something that
needs to be specifically handled during the merging process and it
erroneously wasn't processed at all before. This commit fixes that and
fixes a panic along the way. Stability is now inherited if possible but
generates an error if two packages disagree.

Closes #1666