Gate component model functions with multiple returns (#1670)
This commit is an implementation of WebAssembly/component-model#368 in
this repository. The intention here is along the lines of a few other
recent restrictions which is to avoid outright removing support and
instead stage the changes a bit. To that end a new feature flag is added
and multi-return functions are now gated behind this feature flag. Tests
are updated to avoid using it and otherwise the actual removal of
supporting code is deferred to a later date.