fix(wit-parser): improve checking for stable feature gates
This commit introduces some extra checking for stabilized feature
gates (i.e. `@since` & in-code `Stability::Stable`) with a version
and/or feature specified.

There are two primary changes:
- Ensure that packages containing a `@since` annotation must have a
version specified
- Referring to a future unreleased package version in a `@since`
annotation is an error

In the past `Stability::Stable` feature gates were simply treated as
`Unknown`, which hides the information necessary for downstream
crates/consumers to use (and created the question around whether
referring to future package versions is valid).

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>